[added] shouldHandlerRender hook
This hook is used by the router to determine whether or not
a route handler component should render. It receives one argument,
the current route props, and returns a boolean value.

On the client, this hook can be used to prevent a component from
rendering until it has all the props it needs. On the server, we
will use this hook to determine when the entire route tree is ready
to be rendered.